QPD-1376# moved wavelengths validation to tag class
