+++ quartic_sdk/core/entities/tag.py
@@ -4,7 +4,7 @@ The given file contains the class to refer to the tag entity
 from quartic_sdk.core.entities.base import Base
 import quartic_sdk.utilities.constants as Constants
 from quartic_sdk.core.iterators.tag_data_iterator import TagDataIterator
-from quartic_sdk.utilities.exceptions import IncorrectTagParameterException
+from quartic_sdk.utilities.exceptions import IncorrectTagParameterException,IncorrectWavelengthParamException
 
 
 class Tag(Base):
@@ -20,6 +20,24 @@ class Tag(Base):
         "tag_process_type": Constants.PROCESS_VARIABLE_TYPES,
         "category": Constants.INTELLIGENCE_CATEGORIES
     }
+    @staticmethod
+    def raise_exception_for_wavelegths(wavelengths):
+        """
+        Validate wavelengths passed for a spectral tag. Schema as following
+        {"wavelengths" : ['1460000.0','1460001.0]}
+        :param wavelengths: dict containing key as 'wavelengths' and value as list of wavelengths
+        """
+        assert isinstance(wavelengths, dict), "Wavelengths must be a dict"
+        
+        if not wavelengths.get('wavelengths'):
+            raise IncorrectWavelengthParamException(
+                'Invalid Wavelengths: "wavelengths" key required in dict'
+            )
+        elif not isinstance(wavelengths.get("wavelengths"), list):
+            raise IncorrectWavelengthParamException(
+                'Invalid Wavelengths: Wavelength values must be passed in a list '
+            )
+
 
     def __repr__(self):
         """
@@ -72,6 +90,8 @@ class Tag(Base):
         from quartic_sdk.core.entity_helpers.entity_list import EntityList       
         if wavelengths and self.tag_data_type !=  Constants.TAG_DATA_TYPES[6]:
             raise IncorrectTagParameterException( "Invalid parameters : Wavelengths are only supported with spectral tag type")                    
+        if wavelengths:
+            Tag.raise_exception_for_wavelegths(wavelengths)
         return TagDataIterator.create_tag_data_iterator(
             EntityList(
                 Constants.TAG_ENTITY,
